better sleep functionality
works more like we intended.
-you can only sleep at night
-sleeping changes day to morning
-day count almost works properly
-bug where sleeping wreaks havoc on fatigue & stamina but we are removing those anyway most likely

+++ Beach Vacation/Assets/Scripts/PlayerVitals.cs
@@ -29,8 +29,8 @@ public class PlayerVitals : MonoBehaviour
     
     
     public Slider fatigueSlider;
-    public int maxFatigue;
-    public int fatigueFallRate;
+    public float maxFatigue;
+    public float fatigueFallRate;
 
     public bool fatStage1 = true;
     public bool fatStage2 = true;
